refactor(runtime): split JsObject partials
Separate JsObject's core object operations, shape/value storage, and inline descriptor metadata without changing behavior.
