+++ src/JavaScriptRuntime/JsObject.cs
@@ -1,163 +1,7 @@
 using System;
 using System.Collections;
-using System.Diagnostics;
-using System.Linq;
 
 namespace JavaScriptRuntime;
-
-internal sealed class JsShape
-{
-    internal const int DictionaryPromotionThreshold = 2;
-
-    private enum PropertyNameStorage
-    {
-        Interned,
-        Direct
-    }
-
-    private static readonly ThreadLocal<JsShape> _empty = new (() => new JsShape());
-
-    public static JsShape Empty
-    {
-        get => _empty.Value!;
-    }
-
-    private readonly string[] _propertyNames;
-
-    private readonly Dictionary<string, int>? _slotLookup;
-
-    private Dictionary<string, WeakReference<JsShape>>? _transitions;
-
-    public JsShape()
-    {
-        _propertyNames = System.Array.Empty<string>();
-    }
-
-    internal ReadOnlySpan<string> PropertyNamesInSlotOrder => _propertyNames;
-
-    internal int PropertyCount => _propertyNames.Length;
-
-    internal bool UsesDictionaryLookup => _slotLookup != null;
-
-    internal string GetPropertyNameAtSlot(int slot)
-        => _propertyNames[slot];
-
-    internal IEnumerable<string> EnumeratePropertyNamesSnapshot()
-    {
-        var propertyNames = _propertyNames;
-        for (var slot = 0; slot < propertyNames.Length; slot++)
-        {
-            yield return propertyNames[slot];
-        }
-    }
-
-    internal bool HasTransitionCache => _transitions != null;
-
-    internal int TransitionCacheCount => _transitions?.Count ?? 0;
-
-    private JsShape(string newPropertyName, JsShape parent, PropertyNameStorage propertyNameStorage)
-    {
-        var storedPropertyName = propertyNameStorage == PropertyNameStorage.Interned
-            ? string.Intern(newPropertyName)
-            : newPropertyName;
-
-        var parentPropertyNames = parent._propertyNames;
-        _propertyNames = new string[parentPropertyNames.Length + 1];
-        System.Array.Copy(parentPropertyNames, _propertyNames, parentPropertyNames.Length);
-        _propertyNames[parentPropertyNames.Length] = storedPropertyName;
-
-        if (_propertyNames.Length > DictionaryPromotionThreshold)
-        {
-            _slotLookup = parent._slotLookup != null
-                ? new Dictionary<string, int>(parent._slotLookup, StringComparer.Ordinal)
-                : CreateSlotLookup(parentPropertyNames);
-            _slotLookup[storedPropertyName] = parentPropertyNames.Length;
-        }
-    }
-
-    private JsShape(string deadPropertyName, JsShape parent, bool delete)
-    {
-        var deadSlot = parent.GetSlot(deadPropertyName);
-        if (deadSlot < 0)
-        {
-            throw new InvalidOperationException($"Cannot remove missing shape property '{deadPropertyName}'.");
-        }
-
-        var parentPropertyNames = parent._propertyNames;
-        _propertyNames = new string[parentPropertyNames.Length - 1];
-        if (deadSlot > 0)
-        {
-            System.Array.Copy(parentPropertyNames, 0, _propertyNames, 0, deadSlot);
-        }
-        if (deadSlot < _propertyNames.Length)
-        {
-            System.Array.Copy(
-                parentPropertyNames,
-                deadSlot + 1,
-                _propertyNames,
-                deadSlot,
-                _propertyNames.Length - deadSlot);
-        }
-
-        if (_propertyNames.Length > DictionaryPromotionThreshold)
-        {
-            _slotLookup = CreateSlotLookup(_propertyNames);
-        }
-    }
-
-    private static Dictionary<string, int> CreateSlotLookup(ReadOnlySpan<string> propertyNames)
-    {
-        var slots = new Dictionary<string, int>(propertyNames.Length, StringComparer.Ordinal);
-        for (var slot = 0; slot < propertyNames.Length; slot++)
-        {
-            slots[propertyNames[slot]] = slot;
-        }
-        return slots;
-    }
-
-    public JsShape TransitionTo(string newPropertyName)
-    {
-        if (_transitions != null && _transitions.TryGetValue(newPropertyName, out var weakRef) && weakRef.TryGetTarget(out var existingShape))
-        {
-            return existingShape;
-        }
-        var newShape = new JsShape(newPropertyName, this, PropertyNameStorage.Interned);
-        var transitions = _transitions ??= new Dictionary<string, WeakReference<JsShape>>();
-        transitions[newPropertyName] = new WeakReference<JsShape>(newShape);
-        return newShape;
-    }
-
-    /// <summary>
-    /// Adds a property name without publishing it to the shared shape-transition cache.
-    /// This is for objects populated from untrusted input where neither global string
-    /// interning nor long-lived transition-cache keys are appropriate.
-    /// </summary>
-    public JsShape TransitionToUncached(string newPropertyName)
-        => new JsShape(newPropertyName, this, PropertyNameStorage.Direct);
-
-    public JsShape TransitionAway(string deadPropertyName)
-    {
-        return new JsShape(deadPropertyName, this, true);
-    }
-
-    public int GetSlot(string propertyName)
-    {
-        if (_slotLookup != null)
-        {
-            return _slotLookup.TryGetValue(propertyName, out var slot) ? slot : -1;
-        }
-
-        var propertyNames = _propertyNames;
-        for (var slot = 0; slot < propertyNames.Length; slot++)
-        {
-            if (string.Equals(propertyNames[slot], propertyName, StringComparison.Ordinal))
-            {
-                return slot;
-            }
-        }
-        return -1;
-    }
-}
 
 /// <summary>
 /// Marks a <see cref="JsObject"/> subclass that overrides ECMAScript internal
@@ -183,279 +27,8 @@ internal interface IExoticJsObject
 /// for object literal property initialization to avoid the <c>box</c> instruction.
 /// </para>
 /// </summary>
-public class JsObject : IDictionary<string, object?>
+public partial class JsObject : IDictionary<string, object?>
 {
-    [Flags]
-    private enum JsSlotDescriptorFlags : byte
-    {
-        None = 0,
-        HasMetadata = 1,
-        IsAccessor = 2,
-        Writable = 4,
-        Enumerable = 8,
-        Configurable = 16
-    }
-
-    private struct JsAccessorPair
-    {
-        public object? Get;
-        public object? Set;
-    }
-
-    private sealed class JsObjectDescriptorState
-    {
-        public required byte[] Flags;
-        public JsAccessorPair[]? Accessors;
-        public Dictionary<string, JsPropertyDescriptor>? ExoticOverrides;
-        public HashSet<string>? DeletedLazyClassMethods;
-        public bool HasSharedIntrinsicBaseline;
-    }
-
-    private JsValue[] _properties = System.Array.Empty<JsValue>();
-
-    private JsShape _shape = JsShape.Empty;
-
-    private JsObjectDescriptorState? _descriptorState;
-
-    private object? _prototype;
-
-    private readonly bool _cacheShapeTransitions;
-
-    // Sticky between ordinary mutations, but reset by Clear. While false, every
-    // shape slot is an implicit writable/enumerable/configurable data descriptor.
-    private bool _hasNonDataDescriptors;
-
-    /// <summary>
-    /// True when own reads can no longer be answered from the property dictionary
-    /// alone (the object has accessors, deleted tombstones, or attribute-bearing
-    /// descriptors). It remains sticky until the object is cleared.
-    /// </summary>
-    internal bool HasNonDataDescriptors => _hasNonDataDescriptors;
-
-    internal void MarkNonDataDescriptors() => _hasNonDataDescriptors = true;
-
-    internal bool HasInlineDescriptorState => _descriptorState is not null;
-
-    internal bool HasSharedIntrinsicBaseline
-        => _descriptorState?.HasSharedIntrinsicBaseline == true;
-
-    internal bool HasInlineDescriptors
-        => _shape.PropertyCount != 0
-            || _descriptorState?.ExoticOverrides?.Count > 0;
-
-    internal bool HasInlineExoticDescriptors
-        => _descriptorState?.ExoticOverrides?.Count > 0;
-
-    internal virtual bool UsesInlineExoticDescriptorStorage(string key)
-        => false;
-
-    internal void MarkSharedIntrinsicBaseline()
-    {
-        var state = EnsureDescriptorState();
-        state.HasSharedIntrinsicBaseline = true;
-        MarkNonDataDescriptors();
-        AssertInlineInvariants();
-    }
-
-    internal bool GetInlineOwnDescriptor(string key, out JsPropertyDescriptor descriptor)
-    {
-        var slot = _shape.GetSlot(key);
-        if (slot < 0)
-        {
-            descriptor = default;
-            return false;
-        }
-
-        var flags = _descriptorState is null
-            ? JsSlotDescriptorFlags.None
-            : (JsSlotDescriptorFlags)_descriptorState.Flags[slot];
-        if (flags == JsSlotDescriptorFlags.None)
-        {
-            descriptor = CreateDefaultDataDescriptor(_properties[slot]);
-            return true;
-        }
-
-        if ((flags & JsSlotDescriptorFlags.IsAccessor) != 0)
-        {
-            Debug.Assert(_descriptorState?.Accessors is not null);
-            var accessor = _descriptorState!.Accessors![slot];
-            descriptor = new JsPropertyDescriptor
-            {
-                Kind = JsPropertyDescriptorKind.Accessor,
-                Get = accessor.Get,
-                Set = accessor.Set,
-                Enumerable = (flags & JsSlotDescriptorFlags.Enumerable) != 0,
-                Configurable = (flags & JsSlotDescriptorFlags.Configurable) != 0
-            };
-            return true;
-        }
-
-        descriptor = new JsPropertyDescriptor
-        {
-            Kind = JsPropertyDescriptorKind.Data,
-            Value = _properties[slot].ToObject(),
-            Writable = (flags & JsSlotDescriptorFlags.Writable) != 0,
-            Enumerable = (flags & JsSlotDescriptorFlags.Enumerable) != 0,
-            Configurable = (flags & JsSlotDescriptorFlags.Configurable) != 0
-        };
-        return true;
-    }
-
-    internal void DefineInlineOwnDescriptor(string key, JsPropertyDescriptor descriptor)
-    {
-        var slot = EnsurePropertySlot(key);
-        if (descriptor.Kind == JsPropertyDescriptorKind.Data)
-        {
-            _properties[slot] = JsValue.FromObject(descriptor.Value);
-            if (IsDefaultDataDescriptor(descriptor))
-            {
-                ClearSlotMetadata(slot);
-                AssertInlineInvariants();
-                return;
-            }
-
-            var state = EnsureDescriptorState();
-            state.Flags[slot] = (byte)(
-                JsSlotDescriptorFlags.HasMetadata
-                | (descriptor.Writable ? JsSlotDescriptorFlags.Writable : 0)
-                | (descriptor.Enumerable ? JsSlotDescriptorFlags.Enumerable : 0)
-                | (descriptor.Configurable ? JsSlotDescriptorFlags.Configurable : 0));
-            if (state.Accessors is not null)
-            {
-                state.Accessors[slot] = default;
-            }
-        }
-        else
-        {
-            _properties[slot] = JsValue.Undefined;
-            var state = EnsureDescriptorState();
-            state.Accessors ??= new JsAccessorPair[_shape.PropertyCount];
-            state.Flags[slot] = (byte)(
-                JsSlotDescriptorFlags.HasMetadata
-                | JsSlotDescriptorFlags.IsAccessor
-                | (descriptor.Enumerable ? JsSlotDescriptorFlags.Enumerable : 0)
-                | (descriptor.Configurable ? JsSlotDescriptorFlags.Configurable : 0));
-            state.Accessors[slot] = new JsAccessorPair
-            {
-                Get = descriptor.Get,
-                Set = descriptor.Set
-            };
-        }
-
-        MarkNonDataDescriptors();
-        AssertInlineInvariants();
-    }
-
-    internal bool DeleteInlineOwnDescriptor(string key)
-    {
-        var slot = _shape.GetSlot(key);
-        if (slot < 0)
-        {
-            return false;
-        }
-
-        var newProperties = new JsValue[_properties.Length - 1];
-        CopyAroundRemovedSlot(_properties, newProperties, slot);
-
-        if (_descriptorState is { } state)
-        {
-            var newFlags = new byte[state.Flags.Length - 1];
-            CopyAroundRemovedSlot(state.Flags, newFlags, slot);
-            state.Flags = newFlags;
-
-            if (state.Accessors is not null)
-            {
-                var newAccessors = new JsAccessorPair[state.Accessors.Length - 1];
-                CopyAroundRemovedSlot(state.Accessors, newAccessors, slot);
-                state.Accessors = newAccessors;
-            }
-        }
-
-        _shape = _shape.TransitionAway(key);
-        _properties = newProperties;
-        AssertInlineInvariants();
-        return true;
-    }
-
-    internal IEnumerable<string> GetInlineOwnDescriptorKeys()
-        => _shape.EnumeratePropertyNamesSnapshot();
-
-    internal IEnumerable<string> GetInlineExoticDescriptorKeys()
-        => _descriptorState?.ExoticOverrides?.Keys.ToArray()
-            ?? System.Array.Empty<string>();
-
-    internal bool GetInlineExoticOwnDescriptor(
-        string key,
-        out JsPropertyDescriptor descriptor)
-    {
-        if (_descriptorState?.ExoticOverrides is { } overrides
-            && overrides.TryGetValue(key, out descriptor))
-        {
-            return true;
-        }
-
-        descriptor = default;
-        return false;
-    }
-
-    internal void DefineInlineExoticOwnDescriptor(
-        string key,
-        JsPropertyDescriptor descriptor)
-    {
-        var state = EnsureDescriptorState();
-        state.ExoticOverrides ??= new Dictionary<string, JsPropertyDescriptor>(StringComparer.Ordinal);
-        state.ExoticOverrides[key] = descriptor;
-        MarkNonDataDescriptors();
-        AssertInlineInvariants();
-    }
-
-    internal bool DeleteInlineExoticOwnDescriptor(string key)
-    {
-        var deleted = _descriptorState?.ExoticOverrides?.Remove(key) == true;
-        AssertInlineInvariants();
-        return deleted;
-    }
-
-    internal bool IsInlineLazyClassMethodDeleted(string key)
-        => _descriptorState?.DeletedLazyClassMethods?.Contains(key) == true;
-
-    internal void MarkInlineLazyClassMethodDeleted(string key)
-    {
-        var state = EnsureDescriptorState();
-        state.DeletedLazyClassMethods ??= new HashSet<string>(StringComparer.Ordinal);
-        state.DeletedLazyClassMethods.Add(key);
-        MarkNonDataDescriptors();
-        AssertInlineInvariants();
-    }
-
-    internal void ResetInlineDescriptorState()
-    {
-        Debug.Assert(!HasSharedIntrinsicBaseline);
-        _descriptorState = null;
-        _hasNonDataDescriptors = false;
-        AssertInlineInvariants();
-    }
-
-    internal bool TryGetInlinePrototype(out object? prototype)
-        => (prototype = _prototype) is not null;
-
-    internal void SetInlinePrototype(object? prototype) => _prototype = prototype;
-
-    /// <summary>Creates an ordinary object using the shared shape-transition cache.</summary>
-    public JsObject()
-    {
-        _cacheShapeTransitions = true;
-    }
-
-    /// <summary>
-    /// Creates an ordinary object whose property names are appended without using shared
-    /// shape transitions. Intended for records populated from untrusted property keys.
-    /// </summary>
-    internal JsObject(bool cacheShapeTransitions)
-    {
-        _cacheShapeTransitions = cacheShapeTransitions;
-    }
-
     // -------------------------------------------------------------------------
     // ECMAScript internal object-operation hooks
     // -------------------------------------------------------------------------
@@ -845,155 +418,4 @@ public class JsObject : IDictionary<string, object?>
     }
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
-
-    // -------------------------------------------------------------------------
-    // Helpers
-    // -------------------------------------------------------------------------
-    private void SetValue(string key, JsValue value)
-    {
-        if (HasSharedIntrinsicBaseline && !PropertyDescriptorStore.IsIntrinsicInitialization)
-        {
-            PropertyDescriptorStore.DefineOrUpdate(
-                this,
-                key,
-                CreateDefaultDataDescriptor(value));
-            return;
-        }
-
-        var slot = EnsurePropertySlot(key);
-        _properties[slot] = value;
-        ClearSlotMetadata(slot);
-        AssertInlineInvariants();
-    }
-
-    private JsValue GetValue(string key)
-    {
-        var slot = _shape.GetSlot(key);
-        if (slot == -1)
-            throw new KeyNotFoundException($"Key '{key}' not found.");
-        return _properties[slot];
-    }
-
-    private bool TryGetJsValue(string key, out JsValue value)
-    {
-        var slot = _shape.GetSlot(key);
-        if (slot == -1)
-        {
-            value = JsValue.Undefined;
-            return false;
-        }
-        value = _properties[slot];
-        return true;
-    }
-
-    private int EnsurePropertySlot(string key)
-    {
-        var slot = _shape.GetSlot(key);
-        if (slot >= 0)
-        {
-            return slot;
-        }
-
-        _shape = _cacheShapeTransitions
-            ? _shape.TransitionTo(key)
-            : _shape.TransitionToUncached(key);
-        slot = _shape.GetSlot(key);
-
-        var newProperties = new JsValue[_properties.Length + 1];
-        System.Array.Copy(_properties, newProperties, _properties.Length);
-        _properties = newProperties;
-
-        if (_descriptorState is { } state)
-        {
-            System.Array.Resize(ref state.Flags, _shape.PropertyCount);
-            if (state.Accessors is not null)
-            {
-                System.Array.Resize(ref state.Accessors, _shape.PropertyCount);
-            }
-        }
-
-        AssertInlineInvariants();
-        return slot;
-    }
-
-    private JsObjectDescriptorState EnsureDescriptorState()
-        => _descriptorState ??= new JsObjectDescriptorState
-        {
-            Flags = new byte[_shape.PropertyCount]
-        };
-
-    private void ClearSlotMetadata(int slot)
-    {
-        if (_descriptorState is not { } state)
-        {
-            return;
-        }
-
-        state.Flags[slot] = 0;
-        if (state.Accessors is not null)
-        {
-            state.Accessors[slot] = default;
-        }
-    }
-
-    private static bool IsDefaultDataDescriptor(JsPropertyDescriptor descriptor)
-        => descriptor.Kind == JsPropertyDescriptorKind.Data
-            && descriptor.Writable
-            && descriptor.Enumerable
-            && descriptor.Configurable;
-
-    private static JsPropertyDescriptor CreateDefaultDataDescriptor(JsValue value)
-        => new()
-        {
-            Kind = JsPropertyDescriptorKind.Data,
-            Value = value.ToObject(),
-            Writable = true,
-            Enumerable = true,
-            Configurable = true
-        };
-
-    private static void CopyAroundRemovedSlot<T>(T[] source, T[] destination, int slot)
-    {
-        if (slot > 0)
-        {
-            System.Array.Copy(source, 0, destination, 0, slot);
-        }
-
-        if (slot < destination.Length)
-        {
-            System.Array.Copy(source, slot + 1, destination, slot, destination.Length - slot);
-        }
-    }
-
-    [Conditional("DEBUG")]
-    private void AssertInlineInvariants()
-    {
-        Debug.Assert(_properties.Length == _shape.PropertyCount);
-        if (_descriptorState is not { } state)
-        {
-            return;
-        }
-
-        Debug.Assert(state.Flags.Length == _shape.PropertyCount);
-        Debug.Assert(state.Accessors is null || state.Accessors.Length == _shape.PropertyCount);
-        for (var slot = 0; slot < state.Flags.Length; slot++)
-        {
-            var flags = (JsSlotDescriptorFlags)state.Flags[slot];
-            Debug.Assert(
-                flags == JsSlotDescriptorFlags.None
-                || (flags & JsSlotDescriptorFlags.HasMetadata) != 0);
-            Debug.Assert(
-                (flags & JsSlotDescriptorFlags.IsAccessor) == 0
-                || state.Accessors is not null);
-        }
-
-        if (state.ExoticOverrides is not null)
-        {
-            foreach (var key in state.ExoticOverrides.Keys)
-            {
-                Debug.Assert(UsesInlineExoticDescriptorStorage(key));
-                Debug.Assert(_shape.GetSlot(key) < 0);
-            }
-        }
-    }
 }
